Revert switch to native Sire indexing.
While the indexing itself appears to work, there are issues during
GCMC simulation. My hunch is that the issue is in a mismatch in
the order in which molecules are added to an OpenMM system versus
the order in which the are returned for indexing, or added to the
Sire system.